Update upsert code example
