+++ source/sdk/kotlin/realm-database/crud/update.txt
@@ -78,7 +78,7 @@ To upsert into a realm:
    existing object. If no object exists that matches your filter, realm
    inserts a new object.
 
-.. literalinclude:: /examples/generated/kotlin/CRUDTest.snippet.upsert-an-object.kt
+.. literalinclude:: /examples/generated/kotlin/UpdateTest.snippet.upsert-an-object.kt
    :language: kotlin
 
 .. _kotlin-update-collection:
